Add support for building eBook with Docker
This change adds support for building the eBook with Docker. Supporting Docker means
that users who want to build the eBook do not have to install Ruby/Bundler/Calibre in their
environments.

I'm a Docker n00b, so if there's a better way to handle file permissions on the output, I'm all
ears.

Build tested on Linux. Outputs were fine.

There is also a change to the pandoc outputs. It seems the `+smart` extension is not supported for epub/html[0]
in their latest versions.

[0] - https://pandoc.org/MANUAL.html#extension-smart