Update README command to use Docker Hub
The instructions had a reference to an image hosted on GitHub Container Registry, however that image either doesn't exist or is not publicly accessible. It is also not accessible when authenticated with a GitHub Personal Access Token. The image is available on Docker Hub so this commit updates the command to use Docker Hub as the registry.